Performed refactoring.

+++ src/algorithms/2d/image.cpp
@@ -1,6 +1,6 @@
 #include "image.h"
 
-image::image(unsigned long width, unsigned long height)
+image::image(size_t width, size_t height)
 {
     this->_width = width;
     this->_height = height;
